+++ packages/docusaurus-remark-plugin-npm2yarn/package.json
@@ -8,8 +8,7 @@
     },
     "license": "MIT",
     "dependencies": {
-      "npm-to-yarn": "^1.0.0-2",
-      "@docusaurus/mdx-loader": "2.0.0-alpha.65"
+      "npm-to-yarn": "^1.0.1"
     },
     "devDependencies": {
       "remark": "^12.0.0",
