+++ packages/docusaurus-remark-plugin-npm2yarn/src/__tests__/fixtures/installation.md
